Create P0 Stubs
This commit creates the stubs to be used in the first phase of
development. It also wires up databinding and updates some libraries
that were being used.

Not included are any class definitions or dependencies (execpt for the
aforementioned view model data binding) or any utility classes as I am
unable to predict where those may spring up.

If NASServer becomes too large of a class, it may be useful to split it
up, but for now I made all NAS operations go through it.